handle alarm and pause syscalls + significant signal overhaul
Summary:
When trying to land the pipes/polling change (D40383191), the test rr_sigprocmask_in_syscallbuf_sighandler was a blocker. It uses alarm, and that was not supported at all.  It "worked" before but only by accident -- it delivered its nondeterministic wakeup at a consistent point (because nothing was racing).

This diff intercepts and determinizes alarm() calls.  To do so, it involves quite a bit of plumbing and new functions:

- scheduler notifies guest upon responding whether it was a normal resume or signal-interrupted,
- guest delegates alarm syscalls to the scheduler (register_alarm)
- the scheduler TimedEvents includes upcoming alarms as well as
- add a new alarm test case

And while doing that, we need to ensure various properties for correct semantics:

- only one alarm per pid can be set at once
- when we deliver signals, we have a policy to pick which thread in the process group we target (not 100% complete)
- when clock_nanosleep is interrupted, its rem argument is set with the remaining time
- when a thread is interrupted from blocking by a signal, it must become runnable (go back in the runqueue)

Remaining TODOs:
 - add proper EINTR logic to other blocking syscalls (T137247025)

Reviewed By: jasonwhite

Differential Revision: D41071846

fbshipit-source-id: 971d5b4fbad0360b97365d07218462886d6e12aa