bugfix to schedule normalization
Summary:
After we minimize, we "knockout one" and expect it to fail. That had a bug in it, where there were still "priority=1000" entries that are meaningless, so knocking them out doesn't help.

The fix in this diff is, during schedule normalization, we clean off any of these meaningless default-priority entries.

Reviewed By: VladimirMakaev

Differential Revision: D41081330

fbshipit-source-id: 3833f51662474de932dac9ac28e258b00f6698a2