retry knockouts until we have a nearby-non-matching preemption record
Summary:
Analyze would panic if knocking out an intervention left by the minimize phase left the run still matching the criteria (e.g. crashing).

However, the minimize phase is single-pass, and I'm not sure this is strictly sound.  (I.e. we'd need a second greedy pass to try to re-knockout everything remaining at the end of minimize if we really wanted to guarantee this property.)

Rather than adding that second verification pass, we simply repeat knockouts until we find a pair of schedules differing by one intervention, with a match/dont relationship.

Reviewed By: VladimirMakaev

Differential Revision: D41082639

fbshipit-source-id: 19116878ee167e48b05c3440e2a150b6b264a512