experimental --stacktrace-signals flag
Summary:
This is for signaling the guest at particular points in the execution.

One application i passing SIGQUIT to get a coredump.

Another application is printing stack traces externally via a debugger. That particular application is unfinshed right now.  It needs to be combined with a feature at the hermit analyze level to launch gdb with an init file something like:

```
while(1)
  continue
    bt
end
```

But, unfortunately the SIGINT events are not currently reaching gdb correctly.  Instead I'm seeing this error:

```
Program terminated with signal SIG116, Real-time event 116.
The program no longer exists.
```

Nevertheless

Reviewed By: VladimirMakaev

Differential Revision: D41317754

fbshipit-source-id: fcf8b4112f221022bd3344c08f24dafba92872ca