Count and summarize DESYNC events.
Summary:
When replaying from raw schedule events, we print various kinds of DESYNC messages to the log, and some of the test scripts check the logs for these messages.
In this diff, we directly count these events in the state.  So we can tally them up at the end.

We could also have the hermit `run` function return these summary statistics directly to a caller, for when it's used as a library.  But that isn't done yet.

Reviewed By: VladimirMakaev

Differential Revision: D41425081

fbshipit-source-id: 3dd36b060180cbd57f144d7d7a3e61e5233a76f5