dont resolve env vars during parsing
Summary:
A parsing function shouldn't be performing side effects, like in this case reading from the environment.
This diff defers

But this does change how RunOpts prints.  It round-trips better, because `--env X` remains the same when printed.  But it isn't as explicit as the prior version which fully resolves the argument to `--env X=Y`.

Reviewed By: VladimirMakaev

Differential Revision: D41428092

fbshipit-source-id: 52deaa6d8e64d24d0a11c77d9dd84a9dc4557e7b