a regression test regarding clock total order across threads
Summary:
If we have a causal synchronization from thread1 to thread2, we would expect clock measurements that are ordered (by that "happens before" relation) to also be ordered.

Hermit already satisfies this condition.  Because even if thread-local time lags behind global (vector clock) time, whenever we service the clock_gettime syscall, we use the global time, so it is nice and monotonic.

But because I'm changing things with how local/global time relate, I want this test there as regression protection.

Note: this isn't working under trace-replay mode yet, because of a divergence in an extra register-prehook (plus a downstream difference in a sleep(0) vs TraceReplay COMMIT).

Reviewed By: VladimirMakaev

Differential Revision: D40590539

fbshipit-source-id: d710cb2c7fd2b65fb57babf56d8f81014314a375