make --chaos implied by analyze
Summary:
One could hypothetically want to run analyze on a non-chaos hermit-run (varying PRNG only for example), but this is less likely than that someone makes an error by not including --chaos.
The behavior in that case is not great, because `--search` runs forever repeatedly, and pointlessly, exploring the same deterministic execution.

Therefore, have analyze set chaos to true by default.

Reviewed By: VladimirMakaev

Differential Revision: D41433846

fbshipit-source-id: 455497e84ea1d0918274d26dadd003b5c34dec98