disable preemption recording when replaying schedule events
Summary:
This is a bugfix.  If the chaos preemption/priority recording mechanism is active simulataneously with the schedule-event-trace-replay, the mechanisms can interfere in bad ways. In particular we can hit this assertion failure:

```
thread 'main' panicked at 'assertion failed: `(left == right)`
  left: `8816`,
  right: `9999`', hermetic_infra/hermit/detcore/src/preemptions.rs:623:9
```

That's because the preemption/priority mechanism expects to be in the drivers seat, and see the priority it set last for a thread.  But the trace-replay mechanics are switching priorities between 1 and 9999 as a mechanism of context switching, foregrounding and backgrounding threads.

The simple fix here is to disable the preemption/priority recording while trace replay is active.

This mistake will be easier to avoid, and its absence easier to enforce, when the --record-schedule-to flag and file destination is separated out from the preemption recorder.

Reviewed By: jasonwhite

Differential Revision: D41437114

fbshipit-source-id: 61ef92025b392b3f58b794c564b39248df9029c0