respect branch count during replay - 2/x. Add SignalReceived event to traced SchedEvents
Summary:
When recording execution with signals we currently ignore signal handler. This creates a recorded sequence of "B", "O", "B", "O" loosing the fact that Signal handler fired in the middle.

So to properly handle branch counts while replaying we need to record the fact that a signal handler executed in between

Reviewed By: rrnewton

Differential Revision: D41147141

fbshipit-source-id: 19d0b4a37b91e14a1166d4f1b7990e600e22ab61