respect branch count during replay - 3/x. Add a resource request and trace_schedevent in handle_signal
Summary:
When recording execution with signals we currently ignore signal handler. This creates a recorded sequence of "B", "O", "B", "O" loosing the fact that Signal handler fired in the middle.

So to properly handle branch counts while replaying we need to record the fact that a signal handler executed in between

We also

Reviewed By: rrnewton

Differential Revision: D41069728

fbshipit-source-id: 7257977dd4098585c2cdd898de282c23e3eaf5db