go compatibility: turn sigaction on SIGSTKFLT into noop instead of error
Summary: As stated in the comment, Go registers this signal handler even though the signal should be unused.  Go gets unhappy if we return EINVAL, so we just let it behave as a noop instead.  Such signals should never be delivered to the guest, so whether a handler was changed or not should not matter.

Differential Revision: D41615308

fbshipit-source-id: 40feb3bdaf49dee77716442260142b60f728680b