minor: remove integer overflow
Summary:
Now that hermit is working with ASAN, it would be nice to flip our tests over to that mode.  There are currently about 20 tests that have some problem, and the main thing is just C tests that either leak memory or have overflow, as in the "don't care" pointless work inside the test modified in this diff.

This diff just replaces that meaningless work in the test with something different.

Reviewed By: jasonwhite

Differential Revision: D41622042

fbshipit-source-id: 074ca718a92472b89c3d22231e7423eefdd68939