lay groundwork for thread relative time in SchedEvents
Summary:
I plan to remove starting_micros entirely from DetTime, and only track a notion of local work in that type.

This is a small step in that direction, introducing a way to ask the current DetTime for its local-work-only (sans starting time).
In the next diff we will use that as the "end_time" field of SchedEvent.

Reviewed By: VladimirMakaev

Differential Revision: D41653957

fbshipit-source-id: 00f3d79df009e432fd6a9f06b88c355f28f8215f