refactor update_logical_time_rcbs to not call trace_schedevent
Summary:
In debugging the mem_race test under tracereplay_chaos, it became clear that we were not accurately reproducing the registers DETLOG printout for the prehook of a timer preemption event.

The problem was that update_logical_time_rcbs was IMMEDIATELY doing the side effect of yielding (via trace_schedevent).
This refactor has update_logical_time_rcbs instead returning the events as a pure value, and it defers the side effect till later on.

Reviewed By: jasonwhite

Differential Revision: D41817198

fbshipit-source-id: 3cb302cbf46e66c48f0c91fd1b70debe2b072567