set MAX next timeslice when replaying schedule traces
Summary:
On the mem_race test (tracereplay_chaos mode) it became clear that we were setting a bogus 200_000_000ns preemption timer, which didn't match the original --chaos run anyway.

We depend on the branch-respecting mechanism to set the timeslice when consuming the SchedEvent.  This timeslice set by next_timeslice should be None/MAX.  This diff sets it to MAX.

Reviewed By: VladimirMakaev

Differential Revision: D41825702

fbshipit-source-id: a8851fff53e743f70a7359b88d805e65cc084bfe