eliminate spurious OtherInstructions after precise Branch
Summary:
We use the same prehook logic in the handle_timer_event callback as when handling nondeterministic insturctions, but this includes an unnecessary "OtherInstructions".  This is technically correct, because OtherInstructions is a wildcard that stands for ZERO or more instructions. But is pointless to generate the event in this context, because a branch-based timer will ALWAYS fire right after a branch boundary.

When we change our set_timer_precise to use RcbsAndInstructions, then this invariant will change. In that case we should know precisely how many instructions actually did occur, which we cannot currently encode in the SchedEvent.  We could add a new optional InstructionCount field.  It would be Some(1) for syscalls/cpuid/rdtsc but None (unknown) for Branch events. For OtherInstructions it could be either unknown (0 or more) or a precise number.

You can see the change this diff makes by running a trace-replay verification and seeing Branches as the context switching point instead of alOtherInstructions:

```
$ hermit-verify --verbose trace-replay --chaos --keep-temp-dir --isolate-workdir  -- ~/temp/rustbin_mem_race
...
  4: (dtid7 cnt=9351663 time=105_266_195ns Branch) => (dtid3 end=0x7ffff7d2d1c9 time=11_669_425ns Syscall(clone3, Posthook))
  5: (dtid3 cnt=12405211 time=135_742_005ns Branch) => (dtid7 cnt=5648900 time=161_755_195ns Branch)
```

Reviewed By: VladimirMakaev

Differential Revision: D41841835

fbshipit-source-id: d8ecad18595a5234c7f05b1242910b0177632cc7