overhaul the way runs happen, introduce RunData type
Summary:
Introducing a post-processing step (stripping timestamps) on recorded schedules was proving difficult, because this file had a mess of paths passed around everywhere for logs/preemptions/schedules.

This overhaul attempts to place everything having to do with a single run in its own type.  That type should be in charge of where the different files go, and whether the data is in memory or just on disk.

Like many things in our designs, it has a monotonicity property.  Reading from disk to memory should happen once, with the in-memory fields being write-once immutable, like Ivars.

Differential Revision: D41876234

fbshipit-source-id: fcb787dbf6a9b4ad4fc9c1571eea874b17480b57