flip things and report stack traces from passing run
Summary:
Epecially when we consider the runs that hard-crash in the failing (target) run, it makes sense to report the stack traces from the passing (baseline) run.

As a debugging option, we retain the ability to print the critical events from the target run as well, when --verbose is specified.

Differential Revision: D41918223

fbshipit-source-id: 8185d0c1f6faaeaa86fd6581116b2bb696c2d666