flip on jitter-aware search
Summary:
This is a major change.  Until now, the analyzer has not been reporting any jitter to the search algorithm.  Instead we were basically cruising along what we *thought* was the original bisection path between source and target endpoint runs.

Now, we record, read back, and return the new schedule each time we run, and the search uses it to compute our new position to search from.

Unfortunately, this means the search is now non-monotonic, and edit distance can definitely INCREASE due to jitter.

#closes T137832979

Differential Revision: D41434861

fbshipit-source-id: 5f56abb42eae17bdb8132685ad77ccaae4bb625f