RunData overhaul: push through the target-endpoint step
Summary:
Push the overhaul through the next step in the main() function. This is the step that verifies the target endpoint and converts it to a schedule trace. In particular `save_final_target_sched_events` now returns RunData.

There's a change in behavior here as well (optimization) which is that we don't do the extra run if we don't need it.

Differential Revision: D41891438

fbshipit-source-id: fee03839a796736b10a6e50c51ed34b6e4034d2e