track desyncs at the context switch points specifically
Summary: Refactor to introduce a DesyncStats type.  Within that, track the at_context_switch desyncs.  These are the (very bad) events where we come up to a context switch point, during schedule replay, but we're desynced on the current event, and thus we don't really know if we're following the schedule correctly or not.  In this state pretty much all bets are off, except that you will get SOME valid schedule.

Differential Revision: D42003062

fbshipit-source-id: 58184df52476e5c3c853d194336a10810cfe8b35