minor: change log message to refer to thread-local time
Summary:
Because thread local "times" no longer keep up with global time, I'm trying to make a distinction in the logs where:

(1) if you see a time with the massive epoch  seconds value added (1640995199022558375, or 1_640_995_199.022_558_375s), then that is a global time
(2) if it is a much smaller value, typically sub-second, then it is the thread-local time

The thread-local times still include time "inherited" from the parent.  So they're not (yet?) a pure record of local work (rcbs + syscalls).

Differential Revision: D42005383

fbshipit-source-id: d438d2df6c66fcba5b0e35ff054bd841636e1641