refactor to track actual SchedEvent for each stack to print
Summary:
stacktrace_events works by event index.
But when we're replaying from a recording, this should refer to the event index in that recording.

If any desync occurs, its important that we print the desired event, NOT the event that happens to have that index.

This diff doesn't change the behavior but takes the first step: track the actual SchedEvent in the iterator over events to print.

Differential Revision: D42042568

fbshipit-source-id: 6d0ef4f8c704d5bd480b6ab1ae59b887ffd6fae6