add fields to Report and print execution context w analyze output
Summary:
The report already includes stack traces, but this change prints a window of raw events before/after the critical ones.
In the case of racing syscalls, this makes it a little more visible what the threads are actually doing.

For other operations, it would be much better to gather the stack traces of all these events and print the actual source lines to give an execution history.
(If source files can be found.)

Differential Revision: D42054877

fbshipit-source-id: 97192a0ea68298a537266b86acd7d13724853874