colored 1.9 -> 2.1.0
Summary: A newer version of colored adds support for true color terminal text coloring. The changelog only indicates "Alter Color interface to return Cow<'static, str>" besides the true color support as a "change" in the changelog and I'm not expecting people to really use this. If they do, I'll see it in Sandcastle?

Reviewed By: dtolnay

Differential Revision: D52391497

fbshipit-source-id: c08fffc171f7d5f9b75f96617aa673867ab45312