Fix broken tests after toolchain update
Summary:
These tests could potentially do unaligned pointer reads because the test data is not guaranteed to be aligned.

There is also now a debug assertion in `CString::from_vec_with_nul_unchecked` that ensures the NUL byte is at the end of the `Vec`. Names in dirents may be padded out to keep them aligned.

Reviewed By: dtolnay

Differential Revision: D53597638

fbshipit-source-id: ad6018564d6afd4e8f5de4390d63f64e6c5cadf7